Refactor(error): Domain 객체가 Web 객체를 의존하지 않도록 패키지 구조 변경
