+++ src/main/java/podo/odeego/domain/group/exception/AlreadyParticipatingGroupException.java
@@ -1,7 +1,7 @@
 package podo.odeego.domain.group.exception;
 
-import podo.odeego.web.error.ErrorCode;
-import podo.odeego.web.error.exception.InvalidValueException;
+import podo.odeego.global.error.ErrorCode;
+import podo.odeego.global.error.exception.InvalidValueException;
 
 public class AlreadyParticipatingGroupException extends InvalidValueException {
