+++ src/main/java/podo/odeego/domain/group/exception/GroupAlreadyContainsException.java
@@ -1,7 +1,7 @@
 package podo.odeego.domain.group.exception;
 
-import podo.odeego.web.error.ErrorCode;
-import podo.odeego.web.error.exception.BusinessException;
+import podo.odeego.global.error.ErrorCode;
+import podo.odeego.global.error.exception.BusinessException;
 
 public class GroupAlreadyContainsException extends BusinessException {
