+++ src/main/java/podo/odeego/domain/group/exception/GroupHostAbsentException.java
@@ -1,7 +1,7 @@
 package podo.odeego.domain.group.exception;
 
-import podo.odeego.web.error.ErrorCode;
-import podo.odeego.web.error.exception.InvalidValueException;
+import podo.odeego.global.error.ErrorCode;
+import podo.odeego.global.error.exception.InvalidValueException;
 
 public class GroupHostAbsentException extends InvalidValueException {
