+++ src/main/java/podo/odeego/domain/member/exception/MemberNicknameDuplicatedException.java
@@ -1,7 +1,7 @@
 package podo.odeego.domain.member.exception;
 
-import podo.odeego.web.error.ErrorCode;
-import podo.odeego.web.error.exception.InvalidValueException;
+import podo.odeego.global.error.ErrorCode;
+import podo.odeego.global.error.exception.InvalidValueException;
 
 public class MemberNicknameDuplicatedException extends InvalidValueException {
