+++ src/main/java/podo/odeego/global/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package podo.odeego.web.error;
+package podo.odeego.global.error;
 
 import org.springframework.http.HttpStatus;
