+++ src/main/java/podo/odeego/global/error/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package podo.odeego.web.error;
+package podo.odeego.global.error;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
-import podo.odeego.web.error.exception.BusinessException;
+import podo.odeego.global.error.exception.BusinessException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
